Improved the project structure and installed nodemon to automaticaly connect to the server.
